enriching field meta only if serialisation is required
